Update build-integration-website-hosting.md (#286)
* Update build-integration-website-hosting.md

I ran into an issue where CRA apps or VITE react apps would not work or load correctly with the BrowserRouter from react-router-dom. It took me a considerable amount of time to figure out the cause of the issue and fix it. I hope this is helpful (:

* Update build-integration-website-hosting.md

updated the formatting as requested (: